+++ include/cub3d_typedef.h
@@ -6,7 +6,7 @@
 /*   By: Yoshihiro Kosaka <ykosaka@student.42tok    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/09 13:03:00 by ykosaka           #+#    #+#             */
-/*   Updated: 2023/01/29 20:26:47 by Yoshihiro K      ###   ########.fr       */
+/*   Updated: 2023/02/02 09:42:00 by Yoshihiro K      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,9 +28,9 @@ typedef struct s_coord
 
 typedef struct s_angle
 {
-	int	pitch;
-	int	yaw;
-	int	roll;
+	double	pitch;
+	double	yaw;
+	double	roll;
 }	t_angle;
 
 typedef struct s_color
