+++ src/ft_math.c
@@ -6,7 +6,7 @@
 /*   By: Yoshihiro Kosaka <ykosaka@student.42tok    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/09 13:03:00 by ykosaka           #+#    #+#             */
-/*   Updated: 2023/01/30 13:23:34 by Yoshihiro K      ###   ########.fr       */
+/*   Updated: 2023/02/02 09:44:10 by Yoshihiro K      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,12 +22,12 @@ double	ft_math_deg2rad(int deg)
 	return (M_PI * deg / 180);
 }
 
-int	ft_math_rad2deg(double rad)
+double	ft_math_rad2deg(double rad)
 {
 	return ((int)(rad * 180 / M_PI));
 }
 
-int	ft_math_angle_normalize(int angle)
+double	ft_math_angle_normalize(double angle)
 {
 	if (angle > 180)
 		angle -= 360;
